+++ tests/end_to_end_tests/test_register_dataset_dummy.py
@@ -6,7 +6,7 @@ import pytest
 import sqlalchemy
 import yaml
 from dataregistry import DataRegistry
-from dataregistry.db_basic import SCHEMA_VERSION
+from dataregistry.schema import DEFAULT_SCHEMA_WORKING
 from dataregistry.registrar.dataset_util import get_dataset_status, set_dataset_status
 from dataregistry.registrar.registrar_util import _form_dataset_path
 
@@ -23,7 +23,7 @@ def test_register_dataset_defaults(dummy_file):
 
     # Establish connection to database
     tmp_src_dir, tmp_root_dir = dummy_file
-    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=SCHEMA_VERSION)
+    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=DEFAULT_SCHEMA_WORKING)
 
     # Add entry
     d_id = _insert_dataset_entry(
@@ -50,7 +50,6 @@ def test_register_dataset_defaults(dummy_file):
     assert results["owner_type"][0] == "user"
     assert results["description"][0] == None
     assert results["relative_path"][0] == f"{_NAME}_0.0.1"
-    assert results["version_suffix"][0] == None
     assert results["data_org"][0] == "dummy"
     assert results["execution_id"][0] >= 0
     assert results["dataset_id"][0] >= 0
@@ -88,13 +87,12 @@ def test_register_dataset_manual(dummy_file):
     _OWNER = "test_owner"
     _OWNER_TYPE = "group"
     _REL_PATH = "manual/rel/path"
-    _V_SUFFIX = "test"
     _ACCESS_API = "test_api"
     _IS_OVERWRITABLE = True
 
     # Establish connection to database
     tmp_src_dir, tmp_root_dir = dummy_file
-    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=SCHEMA_VERSION)
+    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=DEFAULT_SCHEMA_WORKING)
 
     # Add entry
     d_id = _insert_dataset_entry(
@@ -105,7 +103,6 @@ def test_register_dataset_manual(dummy_file):
         owner=_OWNER,
         owner_type=_OWNER_TYPE,
         relative_path=_REL_PATH,
-        version_suffix=_V_SUFFIX,
         access_api=_ACCESS_API,
         is_overwritable=_IS_OVERWRITABLE,
     )
@@ -128,7 +125,6 @@ def test_register_dataset_manual(dummy_file):
     assert results["owner_type"][0] == _OWNER_TYPE
     assert results["description"][0] == _DESCRIPTION
     assert results["relative_path"][0] == _REL_PATH
-    assert results["version_suffix"][0] == _V_SUFFIX
     assert results["data_org"][0] == "dummy"
     assert results["execution_id"][0] >= 0
     assert results["dataset_id"][0] >= 0
@@ -155,32 +151,6 @@ def test_register_dataset_manual(dummy_file):
     assert results["replace_iteration"][0] == 0
 
 
-def test_bump_vsuffix(dummy_file):
-    """Should not be able to bump datasets with a version suffix"""
-
-    _NAME = "DESC:datasets:test_bump_vsuffix"
-
-    # Establish connection to database
-    tmp_src_dir, tmp_root_dir = dummy_file
-    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=SCHEMA_VERSION)
-
-    # Add entry
-    d_id = _insert_dataset_entry(
-        datareg,
-        _NAME,
-        "0.0.1",
-        version_suffix="custom_suffix",
-    )
-
-    # Try to bump dataset with version suffix (should fail)
-    with pytest.raises(ValueError, match="Cannot bump"):
-        d_id = _insert_dataset_entry(
-            datareg,
-            _NAME,
-            "major",
-        )
-
-
 @pytest.mark.parametrize(
     "v_type,ans",
     [
@@ -199,7 +169,7 @@ def test_dataset_bumping(dummy_file, v_type, ans):
 
     # Establish connection to database
     tmp_src_dir, tmp_root_dir = dummy_file
-    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=SCHEMA_VERSION)
+    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=DEFAULT_SCHEMA_WORKING)
 
     # Add entry
     d_id = _insert_dataset_entry(
@@ -233,7 +203,7 @@ def test_dataset_owner_types(dummy_file, owner_type):
 
     # Establish connection to database
     tmp_src_dir, tmp_root_dir = dummy_file
-    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=SCHEMA_VERSION)
+    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=DEFAULT_SCHEMA_WORKING)
 
     # Add entry
     d_id = _insert_dataset_entry(
@@ -266,7 +236,7 @@ def test_register_dataset_with_global_owner_set(dummy_file):
     tmp_src_dir, tmp_root_dir = dummy_file
     datareg = DataRegistry(
         root_dir=str(tmp_root_dir),
-        schema=SCHEMA_VERSION,
+        schema=DEFAULT_SCHEMA_WORKING,
         owner="DESC group",
         owner_type="group",
     )
@@ -307,7 +277,7 @@ def test_register_dataset_with_modified_default_execution(dummy_file):
 
     # Establish connection to database
     tmp_src_dir, tmp_root_dir = dummy_file
-    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=SCHEMA_VERSION)
+    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=DEFAULT_SCHEMA_WORKING)
 
     d_id_1 = _insert_dataset_entry(
         datareg,
@@ -380,7 +350,7 @@ def test_dataset_query_return_format(dummy_file, return_format_str, expected_typ
 
     # Establish connection to database
     tmp_src_dir, tmp_root_dir = dummy_file
-    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=SCHEMA_VERSION)
+    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=DEFAULT_SCHEMA_WORKING)
 
     _NAME = f"DESC:datasets:query_return_test_{return_format_str}"
 
@@ -406,7 +376,7 @@ def test_query_all(dummy_file):
 
     # Establish connection to database
     tmp_src_dir, tmp_root_dir = dummy_file
-    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=SCHEMA_VERSION)
+    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=DEFAULT_SCHEMA_WORKING)
 
     # Register a dataset
     d_id_1 = _insert_dataset_entry(
@@ -429,7 +399,7 @@ def test_dataset_bad_name_string(dummy_file, name):
 
     # Establish connection to database
     tmp_src_dir, tmp_root_dir = dummy_file
-    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=SCHEMA_VERSION)
+    datareg = DataRegistry(root_dir=str(tmp_root_dir), schema=DEFAULT_SCHEMA_WORKING)
 
     # Register a dataset
     with pytest.raises(ValueError, match="Cannot have character"):